% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concaveman.R
\name{concaveman}
\alias{concaveman}
\title{A very fast 2D concave hull algorithm}
\usage{
concaveman(x, y = NULL, concavity = 2, lengthThreshold = 0)
}
\arguments{
\item{x, y}{coordinate vectors of points. This can be specified as two vectors x and y, a 2-column
matrix x, a list x with two components, etc.}

\item{concavity}{numeric a relative measure of concavity. 1 results in a relatively detailed shape,
Infinity results in a convex hull. You can use values lower than 1, but they can produce pretty crazy
shapes.}

\item{lengthThreshold}{numeric. when a segment length is under this threshold, it stops being
considered for further detalization. Higher values result in simpler shapes.}
}
\description{
The algorithm is based on ideas from Park and Oh (2012) (see references) and was first implement
}
\examples{
data(points)
hull <- concaveman(points)
plot(points, asp = 1)
lines(hull, lwd = 3, col = "red")
}
\references{
Park, J.-S & Oh, S.-J. (2013). A New Concave Hull Algorithm and Concaveness Measure
for n-dimensional Datasets. Journal of Information Science and Engineering. 29. 379-392.
}
