% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{points}
\alias{points}
\title{Fixtures data}
\format{
A matrix with a 1000 points.
}
\source{
https://github.com/mapbox/concaveman/blob/master/test/fixtures/points-1k.json
}
\usage{
points
}
\description{
This is just a test dataset which comes from the original mapbox library.
}
\keyword{datasets}
